move FieldInfoSection into separate folder
